refactor: Add bankFields route to handle bank field mapping
This commit adds a new route called `bankFields` to handle the retrieval of bank field mappings. The route imports the `prisma` client from the `db.server` module and uses it to query the `bank_fields_map` table based on the provided `bankName` parameter. If the `bankName` is not provided, a 400 error response is returned. The retrieved `fieldMap` is then returned as a JSON response. This change enhances the functionality of the application by providing a way to retrieve and use bank field mappings in the application.