refactor: Update Navbar component to use Link instead of NavLink
This commit updates the Navbar component in the Navbar.tsx file to use the Link component from the @remix-run/react library instead of the NavLink component. This change improves the performance of the navigation links by removing unnecessary active class updates. It also simplifies the code by removing the need for the space-x-reverse class. Overall, this refactor enhances the efficiency and readability of the Navbar component.